Task 3 Complete
Task 3 has been implemented. After a user registers and logs in, they are redirected to the home page. After reaching this homepage, they can search based on ISBN, Title, or Author and the website will return a list of all books that include the query that was searched. An message is also displayed if there are no matches. If the user only inputs a partial title, author name, or isbn, the search tool will return any matching results.